[#58]fix. roomList 컴포넌트 분리
