+++ FE/airbnb/src/pages/ReservePage.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import Map from '../components/map/Map';
 import ReserveHeader from '../components/reserveHeader/ReserveHeader';
 import MapSkeleton from '../components/reservePageSkeleton/MapSkeleton';
-import ReserveRoomList from '../components/reserveRoomList/ReserveRoomList';
+import ReserveRoomsSection from '../components/reserveRoomList/ReserveRoomsSection';
 import { reserveInfoSelector } from '../recoilStore/headerAtom';
 
 interface Props {}
@@ -33,9 +33,7 @@ const ReservePage = ({}: Props) => {
   return (
     <StyledReservePage>
       <ReserveHeader />
-      <Suspense fallback=''>
-        <ReserveRoomList className='rooms' />
-      </Suspense>
+      <ReserveRoomsSection className='rooms' />
       <Suspense fallback={<MapSkeleton />}>
         <Map className='map' />
       </Suspense>
